Better config

+++ lib/senders/sftp.py
@@ -23,7 +23,7 @@ class SFTPClient(BaseNetworkClient):
     configurable = BaseNetworkClient.configurable.copy()
     configurable.update({'basepath': Path, 'filename': RawStr, 'remotepath': Path, 'sftploglevel': int,
                          'knownhosts': Path, 'username': str, 'password': str, 'clientkeys': Path,
-                         'passphrase': None, 'clientversion': RawStr})
+                         'passphrase': None, 'clientversion': RawStr, 'removetmpfiles': bool})
     sender_type = 'SFTP Client'
     sftp_client = None
     sftp = None
@@ -33,7 +33,7 @@ class SFTPClient(BaseNetworkClient):
     def __init__(self, *args, port=asyncssh.connection._DEFAULT_PORT, srcport=asyncssh.connection._DEFAULT_PORT,
                  basepath=settings.DATA_DIR.joinpath('tmp'), filename=None, remotepath=Path('.'),
                  knownhosts: Path=None, username: str=None, password: str=None, sftp_kwargs=None,
-                 clientkeys: Path=(), passphrase: str=None, clientversion:str = (), **kwargs):
+                 clientkeys: Path=(), passphrase: str=None, clientversion:str = (), removetmpfiles = True, **kwargs):
         self.filename = filename
         self.remotepath = remotepath
         self.base_path = basepath
@@ -44,6 +44,7 @@ class SFTPClient(BaseNetworkClient):
         self.sftp_kwargs.update(sftp_kwargs)
         super(SFTPClient, self).__init__(*args, port=port, srcport=srcport, **kwargs)
         self.mode = 'ba' if self.manager.protocol.binary else 'a'
+        self.removetmpfiles = removetmpfiles
 
     async def open_connection(self, **kwargs):
         self.conn, self.sftp_client = await asyncssh.create_connection(
@@ -63,24 +64,29 @@ class SFTPClient(BaseNetworkClient):
         kwargs['remotepath'] = kwargs.get('remotepath', str(real_remote_path))
         await self.sftp.put(str(file_path), **kwargs)
 
+    def get_tmp_file_path(self, filename=None):
+        if not filename:
+            filename = binascii.hexlify(os.urandom(12)).decode('utf-8')
+        return self.base_path.joinpath(filename)
+
     def encode_msg(self, msg_decoded):
         msg_obj = self.manager.protocol.from_decoded(msg_decoded, sender=self.source)
         if self.filename:
             filename = self.filename.format(msg=msg_obj)
         else:
-            filename = binascii.hexlify(os.urandom(12)).decode('utf-8')
-        path = self.base_path.joinpath(filename)
-        self.logger.debug(path)
+            filename = None
         data = msg_obj.encoded
-        self.raw_log.debug(data)
-        with path.open(self.mode) as f:
-            f.write(data)
-        return path
+        return data, filename
 
     async def encode_and_send_msg(self, msg_decoded):
-        file_path = self.encode_msg(msg_decoded)
-        await self.send_msg(file_path)
-        file_path.unlink()
+        data, filename = self.encode_msg(msg_decoded)
+        await self.send_msg(data, filename=filename)
 
-    async def send_data(self, file_path: AnyStr):
+    async def send_data(self, data, filename=None):
+        file_path = self.get_tmp_file_path(filename)
+        self.logger.debug("Using temp path: %s", file_path)
+        with file_path.open(self.mode) as f:
+            f.write(data)
         await self.put(file_path)
+        if self.removetmpfiles:
+            file_path.unlink()
